Run test suite in Firefox via Karma
Since the primary use-case for react-dates is in a browser environment
it seems appropriate to run the test suite in a browser.

I've added configuration to run the test suite in Firefox via Karma
with the necessary adjustments to the Travis config. I've commented
the two places in the Karma config where I'm doing something odd in
order to make things work.