Use `aria-label` instead of visually hidden label
Instead of using a `label` element with CSS to visually hide it, use the
`aria-label` property on the `input` element itself. Having hidden text for
screen readers could cause the label value to be read aloud twice by the screen
reader: first when navigating nodes in the form, and then again when focusing
the input.