[a11y] expose a callback for when the user hits tab or shift tab inside the day picker
For an upcoming refactor, we need to be able to listen for when the user
hits tab inside the day picker. This is to support an a11y change in
which the daypicker will become an immediate sibling of the date input
so the user can just naturally tab into it. The problem is that when the
user hits tab to exit the day picker, we need to close the calendar,
which is currently done inside the DateRangePickerInputController.